various refacto and doc
