+++ doc/widget-app-structure.md
@@ -1,0 +1,12 @@
+# Widget App Structure
+
+
+#### Map Mode
+Map is decompose in what we call `MapMode`. 
+It has different possible states: 
+- `MapMode.overview` : the map with trails markers
+- `MapMode.preview` : adding UI (1) on the map and previsualizing path
+- `MapMode.trail` : when want to discover / follow a specific trail. 
+
+#### Map Widget tree
+```Insert PICTURE here```
